Fix all clippy warnings and remove worthless tests
- Fix 20 clippy warnings: 6 collapsible_if, 3 derivable_impls,
  3 doc_lazy_continuation, 1 unnecessary_cast, 1 needless_borrow,
  1 len_without_is_empty, 1 new_without_default, 1 new_ret_no_self,
  1 empty_line_after_doc_comments, 1 too_many_arguments
- Add #[allow(clippy::too_many_arguments)] for DraggablePanel::begin
- Rename ComputePass::new to ComputePass::builder (new_ret_no_self)
- Add is_empty() to ShadowBuffers (len_without_is_empty)
- Implement Default for LightOrbitSystem (new_without_default)
- Remove orphaned doc comments in renderer.rs and storage_uniform.rs
- Remove 20 worthless tests: 5 PathBuf::join tests (resources),
  3 gui_state tests, 6 frame_context tests, 10 sync wrapper tests,
  3 error trait tests, commented-out LifetimeComponent derive test

+++ katla_gfx/src/particles/buffer.rs
@@ -179,7 +179,7 @@ impl GlobalParticleBuffer {
                 .get_physical_device_properties(context.physical_device)
         }
         .limits
-        .min_storage_buffer_offset_alignment as u64;
+        .min_storage_buffer_offset_alignment;
 
         let layout = ParticleBufferLayout::new(max_particles, alignment);
